Gameクラスを修正
