+++ 06.bowling_object/shot.rb
@@ -1,0 +1,23 @@
+#!/usr/bin/env ruby
+# frozen_string_literal: true
+
+class Shot
+  attr_reader :marks
+
+  def initialize(marks)
+    @marks = marks
+  end
+
+  def to_shots
+    shots = []
+    marks.split(',').each do |m|
+      if m == 'X'
+        shots << 10
+        shots << 0
+      else
+        shots << m.to_i
+      end
+    end
+    shots
+  end
+end
